Add basic javascript functionality and modify html and css
Add basic functions in the javascript file to recognise the pressing of numbers and operators. Numbers are shown through console.log() for now. As you press operators, you get to a new number.
Add "result" as a button in the calculator and the corresponding css. Also Odin Project's logo on its side.